branch

+++ factory_dc_v2_testing.cpp
@@ -37,7 +37,7 @@ int main()
     char type;
     float value;
     float stop_time = 0;
-    float time_step = 1;
+    long double time_step = 1;
     int component_counter = 0; //component counter for figuring out current number of component
     int voltage_source_counter = 0;
     int capacitor_counter = 0;
